test in progeress

+++ BeersApi/Dto/DrinksResponse.cs
@@ -4,7 +4,7 @@
     {
         public Guid DrinksId { get; set; }
         public string Name { get; set; }
-        public int AlcoholRate { get; set; }
+        public double AlcoholRate { get; set; }
         public int Price { get; set; }
         public Guid DrinkTypeId { get; set; }
     }
