+++ BeersApi/Dto/DrinksUpdate.cs
@@ -4,7 +4,7 @@
     {
         public string? Name { get; set; }
         public int? Price { get; set; }
-        public int? AlcoholRate { get; set; }
+        public double? AlcoholRate { get; set; }
         public Guid? DrinkTypeId { get; set; } 
     }
 }
